refactor(rl): unify post-training policy evaluation
Add one eval-policy command for Headless, Viewer, and external Profile workflows. Reuse the training runtime, record run manifests, and consolidate the focused documentation and tests.
