--- docs/source/api_reference/embodichain/embodichain.learning.rl.motion_policy_evaluation.rst
@@ -1,15 +1,0 @@
-embodichain.learning.rl.motion_policy_evaluation
-================================================
-
-.. automodule:: embodichain.learning.rl.motion_policy_evaluation
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-Native Task
------------
-
-.. automodule:: embodichain.learning.rl.motion_policy_evaluation.native_task
-   :members:
-   :undoc-members:
-   :show-inheritance:
