+++ docs/source/api_reference/embodichain/embodichain.learning.rl.policy_evaluation.rst
@@ -1,0 +1,7 @@
+embodichain.learning.rl.policy_evaluation
+=========================================
+
+.. automodule:: embodichain.learning.rl.policy_evaluation
+   :members:
+   :undoc-members:
+   :show-inheritance:
