+++ docs/source/api_reference/embodichain/embodichain.learning.rl.rst
@@ -18,8 +18,7 @@ collection logic, policy/model builders, and training entry points.
       buffer
       collector
       models
-      motion_policy_evaluation
-      runtime
+      policy_evaluation
       train
       utils
