--- docs/source/api_reference/embodichain/embodichain.learning.rl.runtime.rst
@@ -1,29 +1,0 @@
-embodichain.learning.rl.runtime
-===============================
-
-.. automodule:: embodichain.learning.rl.runtime
-
-   .. rubric:: Module Attributes
-
-   .. autosummary::
-
-      GymPolicyFactory
-
-   .. rubric:: Functions
-
-   .. autosummary::
-
-      build_gym_environment
-      build_gym_policy_runtime
-      build_learning_environment
-      build_learning_policy_runtime
-      resolve_config_reference
-      resolve_torch_device
-      seed_policy_runtime
-
-   .. rubric:: Classes
-
-   .. autosummary::
-
-      GymEnvironmentRuntime
-      PolicyRuntime
