+++ embodichain/__main__.py
@@ -82,9 +82,9 @@ COMMANDS = (
         help="Train an RL agent from a JSON or YAML config.",
     ),
     Command(
-        name="eval-motion-policy",
-        target="embodichain.learning.rl.motion_policy_evaluation.cli:cli",
-        help="Visualize a policy through DexSim Motion Policy Kit.",
+        name="eval-policy",
+        target="embodichain.learning.rl.policy_evaluation.cli:cli",
+        help="Evaluate a trained policy in Headless or Viewer mode.",
     ),
     Command(
         name="annotate-grasp",
