+++ embodichain/learning/rl/policy_evaluation/manifest.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 # ----------------------------------------------------------------------------
 
-"""Minimal index connecting a training run to motion-policy evaluation."""
+"""Index a training run for standalone policy evaluation."""
 
 from __future__ import annotations
 
@@ -23,7 +23,6 @@ import shutil
 from collections.abc import Mapping
 from dataclasses import dataclass
 from pathlib import Path
-from types import MappingProxyType
 from typing import Any
 
 __all__ = ["RUN_MANIFEST_NAME", "RunManifest", "write_run_manifest"]
@@ -36,18 +35,13 @@ class RunManifest:
     """Resolved paths from one EmbodiChain training run."""
 
     root: Path
-    motion_profile: str | None
     configs: Mapping[str, Path]
     checkpoints: Mapping[str, Path | None]
 
     def __post_init__(self) -> None:
         object.__setattr__(self, "root", Path(self.root).resolve())
-        object.__setattr__(self, "configs", MappingProxyType(dict(self.configs)))
-        object.__setattr__(
-            self,
-            "checkpoints",
-            MappingProxyType(dict(self.checkpoints)),
-        )
+        object.__setattr__(self, "configs", dict(self.configs))
+        object.__setattr__(self, "checkpoints", dict(self.checkpoints))
 
     @classmethod
     def load(cls, run: str | Path) -> RunManifest:
@@ -73,12 +67,9 @@ class RunManifest:
             "checkpoints",
             allow_none=True,
         )
-        profile = value.get("motion_profile")
-        if profile is not None and not isinstance(profile, str):
-            raise TypeError("Run manifest motion_profile must be a string or null")
-        return cls(root, profile, configs, checkpoints)
+        return cls(root, configs, checkpoints)
 
-    def select_checkpoint(self, requested: str = "best") -> tuple[str, Path]:
+    def select_checkpoint(self, requested: str = "latest") -> tuple[str, Path]:
         """Select ``best`` or ``latest`` and return its resolved path.
 
         Args:
@@ -109,7 +100,6 @@ def write_run_manifest(
     latest_checkpoint: str | Path,
     best_checkpoint: str | Path | None = None,
     gym_config: str | Path | None = None,
-    motion_profile: str | None = None,
 ) -> Path:
     """Snapshot training configs and write the minimal run manifest.
 
@@ -119,7 +109,6 @@ def write_run_manifest(
         latest_checkpoint: Final saved checkpoint.
         best_checkpoint: Best checkpoint when evaluation selected one.
         gym_config: Referenced task config when the trainer uses one.
-        motion_profile: Default Motion Profile for visual evaluation.
 
     Returns:
         Written manifest path.
@@ -139,7 +128,6 @@ def write_run_manifest(
     }
     value: dict[str, Any] = {
         "schema_version": 1,
-        "motion_profile": motion_profile,
         "configs": configs,
         "checkpoints": checkpoints,
     }
